Apply validation follow-ups
Agent-Logs-Url: https://github.com/Vale717171/archive-of-oblivion/sessions/7b41ac62-49ec-4cdc-841b-ea0f13d6fc24

+++ lib/features/settings/app_settings_provider.dart
@@ -81,7 +81,7 @@ class AppSettingsNotifier extends AsyncNotifier<AppSettings> {
       reduceMotion: reduceMotion,
       highContrast: highContrast,
       commandAssist: commandAssist,
-      textScale: textScale == null ? null : textScale.clamp(0.9, 1.4).toDouble(),
+      textScale: textScale == null ? null : textScale.clamp(0.9, 1.4),
       typewriterMillis: typewriterMillis == null ? null : typewriterMillis.clamp(8, 40),
     );
